:tractor: merge Helpers into PyEchartsConfig
